refactor: all tests are in the tests/ directory
